fix(branch): bug in branch creation
Fixed two issues, one the fixture that resets the database was running
everytime hence there was no unique constraint violation because the old
record was wiped. The second issue is that BranchService was setting
the _branch_id in the new Branch instances to a tuple not a string.

To fix the former i add scope="module" to the fixture decorator. See diff
for the latter.

Additionally added extra tests to ensure those issues don't happen again.